+++ graft/coreth/RELEASES.md
@@ -4,11 +4,12 @@
 
 - Removed deprecated flags `coreth-admin-api-enabled`, `coreth-admin-api-dir`, `tx-regossip-frequency`, `tx-lookup-limit`. Use `admin-api-enabled`, `admin-api-dir`, `regossip-frequency`, `transaction-history` instead.
 - Enabled RPC batch limits by default, and configurable with `batch-request-limit` and `batch-max-response-size`.
-- ACP-226:
+- ACP-226 (Granite):
   - Set expected block gas cost to 0 in Granite network upgrade, removing block gas cost requirements for block building.
   - Add `timeMilliseconds` (Unix uint64) timestamp to block header for Granite upgrade.
   - Add `minDelayExcess` (uint64) to block header for Granite upgrade.
   - Add minimum block building delays to conform the block builder to ACP-226 requirements.
+  - Add minimum delay verification.
 - Update go version to 1.24.8
 
 ## [v0.15.3](https://github.com/ava-labs/coreth/releases/tag/v0.15.3)